+++ CHANGELOG.md
@@ -1,10 +1,25 @@
+## 1.5.7
+
+- Add scroller controller as a parameter. Thanks @Faaatman for the PR!
+- Make attachment button margin configurable. Thanks @felixgabler for the PR!
+- Add `avatarBuilder` and `nameBuilder` methods to resolve user updates and render correct avatars and names. Thanks @dariuspo and @felixgabler for the PR!
+- Fix broken text paddings. Thanks @AdrKacz for reporting!
+- Update dependencies
+
 ## 1.5.6
 
+- Add markdown support. Thanks @felixgabler for the PR!
+- Add Arabic localization and RTL support. Thanks @Faaatman for the PR!
+- Increase send button tapable area. Thanks @felixgabler for the PR!
+- Add new `hidden` send button visibility mode. Thanks @fernandobatels for the PR!
+- Add `previewTapOptions` that allow to configure to open link preview when tapped on preview's image or title. Thanks @felixgabler for the PR!
+- Add `dateHeaderBuilder`. Thanks @arsamme for the PR!
 - Add `onMessageVisibilityChanged` handler. Thanks @felixgabler for the PR!
-- Add `receivedMessageBodyLinkTextStyle` and `sentMessageBodyLinkTextStyle` to the theme. Thanks @felixgabler for the PR!
+- Add `receivedMessageBodyBoldTextStyle`, `receivedMessageBodyCodeTextStyle`, `receivedMessageBodyLinkTextStyle` and `sentMessageBodyBoldTextStyle`, `sentMessageBodyCodeTextStyle`, `sentMessageBodyLinkTextStyle` to the theme. Thanks @felixgabler for the PR!
 - Add `sendButtonMargin` to the theme. Thanks @damian-kaczmarek for the PR!
 - Add German localization. Thanks @felixgabler for the PR!
-- Update to Flutter 2.10.2. Requires Dart >= 2.16.0.
+- Fix emoji only horizontal margin. Thanks @munkius for the PR!
+- Update to Flutter 2.10.4. Requires Dart >= 2.16.0.
 
 ## 1.5.5